Convert Floats to Decimals with Monetary Amounts

+++ Recurly/Resources/CouponDiscountPricing.cs
@@ -17,7 +17,7 @@ namespace Recurly.Resources
 
         /// <value>Value of the fixed discount that this coupon applies.</value>
         [JsonProperty("amount")]
-        public float? Amount { get; set; }
+        public decimal? Amount { get; set; }
 
         /// <value>3-letter ISO 4217 currency code.</value>
         [JsonProperty("currency")]
